move device_command.hpp to impl
