+++ tests/tt_metal/tt_metal/dispatch/dispatch_util/test_device_command.cpp
@@ -2,10 +2,10 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include <tt-metalium/device_command.hpp>
 #include "cq_commands.hpp"
 #include "memcpy.hpp"
 #include <gtest/gtest.h>
+#include "tt_metal/impl/dispatch/device_command.hpp"
 #include "tt_metal/impl/dispatch/device_command_calculator.hpp"
 
 namespace tt::tt_metal {
