+++ tt_metal/impl/buffers/dispatch.cpp
@@ -2,10 +2,11 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include <device_command.hpp>
 #include <device.hpp>
 #include "assert.hpp"
 #include "dispatch.hpp"
+#include "tt_metal/impl/dispatch/device_command.hpp"
+
 #include <tt-metalium/command_queue_interface.hpp>
 #include <tt-metalium/dispatch_settings.hpp>
