+++ tt_metal/impl/program/program_command_sequence.hpp
@@ -7,7 +7,7 @@
 #include <memory>
 #include <vector>
 
-#include <tt-metalium/device_command.hpp>
+#include "tt_metal/impl/dispatch/device_command.hpp"
 
 struct CQDispatchWritePackedCmd;
 struct launch_msg_t;
